cli: Another big refactor to break up command structure and allow repo sub command
